feat(app): a few syntax fixes and small refactor
youtubeEvents had some syntax errors that were corrected, these are temporary and will change as the extension changes. CommandManager was modified for demonstration purposes. test.ts had a few syntax issues that were fixed. i modified the ytconfig interface to include a sample property to demonstrate passing values around, i also refactored the forgeNotifications class to use a simpler init function that is modeled after the one in forgedb. in the constructor i added a little test log that logs the hello message that is instantiated at the bottom of the file. i also modified the command manager property and assignment to be a little easier to understand and is modeled after forgedb, i assume this will need to be changed but it works for demonstration purposes.